finalized menu plugin architecture
