--- src/wezel/menu/__init__.py
@@ -1,16 +1,0 @@
-from . import (
-    numpy,
-    about,
-    edit,
-    folder,
-    scipy,
-    view,
-    test,
-    filter,
-    segment,
-    transform,
-    measure,
-    menubar,
-)
-
-
