+++ src/wezel/menubar/__init__.py
@@ -1,0 +1,8 @@
+from . import (
+    about,
+    edit,
+    folder,
+    view,
+)
+
+
