+++ src/wezel/menubar/about.py
