+++ src/wezel/menubar/edit.py
