+++ src/wezel/widgets/log_to_GUI.py
@@ -8,6 +8,11 @@ import traceback
 from PySide2.QtCore import (QObject, QRunnable, QThreadPool, Signal, Slot)
 from PySide2.QtWidgets import (QVBoxLayout, QWidget, QPlainTextEdit)
 
+import wezel
+import time
+import numpy as np
+import os
+
 
 class WorkerSignals(QObject):
       """
@@ -139,4 +144,102 @@ class LoggingWidget(QWidget):
 
     def logFinished(self):
         self.displayLogs.appendPlainText(str(datetime.datetime.now())[:19] + ': ' + "Finished")
+
+
+
+def test_Function(start, end, signals):
+  signals.log.emit("Calculation Started")
+  for n in range(start, end):
+    time.sleep(1)
+    signals.log.emit("Message " + str(n))  
+  return "Result of the calculation"
+
+
+def test_Copy(series_list, signals):
+    signals.log.emit("Copying {} series".format(len(series_list)))      
+    for series in series_list:
+        signals.log.emit('Copying {}'.format(series.label()))
+        series.copy()  
+    return "Copying complete"
+
+
+def test_export(series, path,signals):
+    signals.log.emit("Start exporting series")
+    for s in series:
+        signals.log.emit("Exporting {} series to {}".format(s.label(), path))
+        s.export(path)
+    return "Series export complete"   
+
+
+def test_array_copy(start, end, signals):
+    signals.log.emit("Copying arrays started")      
+    for n in range(start, end):
+        time.sleep(1)
+        first_array=np.arange(n)
+        second_array=np.copy(first_array)
+        signals.log.emit("{} copied".format(first_array)) 
+    return "Array copying complete"
+
+
+def test_file_write(signals):
+    try:
+        if not os.path.exists("C:\\FileDemo\\"):
+            os.makedirs("C:\\FileDemo\\")
+            signals.log.emit("Directory created")
+        with open('C:\\FileDemo\\myfile.txt', 'w+') as f:
+            f.write('Create a new text file!')
+        signals.log.emit("File created") 
+        for n in range(10):
+            time.sleep(1)
+            with open("C:\\FileDemo\\myfile.txt", "a") as f:
+                f.write("\n This text was added using Append.")
+            signals.log.emit("Text appended to file")
+        return "File operations finished"
+    except FileNotFoundError:
+        signals.log.emit("File myfile.txt does not exist")
+        return "Terminated by error"
+
+        
+    
+    
+
+class Test_FileWrite(wezel.gui.Action):
+    def run(self, app):
+        window = wezel.widgets.LoggingWidget(test_file_write)
+        app.addAsSubWindow(window, "Test File Write and Logging to GUI")
+
+
+class Test_LoggingToGUI(wezel.gui.Action):
+    def run(self, app):
+        window = wezel.widgets.LoggingWidget(test_Function, start=1,end=10)
+        app.addAsSubWindow(window, "Test Logging to GUI")
+
+
+class Test_Copy(wezel.gui.Action):
+    def enable(self, app):
+        return app.nr_selected('Series') != 0
+
+    def run(self, app):
+        series_list = app.selected('Series')
+        window = wezel.widgets.LoggingWidget(test_Copy, series_list=series_list)
+        app.addAsSubWindow(window, "Test Logging to GUI while copying DICOM series")
+       
+
+class Test_ArrayCopy(wezel.gui.Action):  
+    def run(self, app):
+        window = wezel.widgets.LoggingWidget(test_array_copy, start=10,end=20)
+        app.addAsSubWindow(window, "Test Logging to GUI while copying numpy arrays")
+
+
+class Test_Export(wezel.gui.Action):
+    """Export selected series"""
+
+    def enable(self, app):
+        return app.nr_selected('Series') != 0
+
+    def run(self, app):
+        series = app.selected('Series')
+        path = app.dialog.directory("Where do you want to export the data?")
+        window = wezel.widgets.LoggingWidget(test_export, series=series, path=path)
+        app.addAsSubWindow(window, "Test Logging to GUI while copying numpy arrays")
